[feat] Data Encryption (#63)
Wow, this was a lot. This adds data encryption for mood logging, surveys, and gap fund, and does so with a new cloud function (the whole function architecture was moved to typescript in the process) and a new auth package for CloudKit. This also came with a bunch of modifications to capacitor-firebase, which hopefully we'll be able to merge in very soon, along with a later PR for getting an authToken on Android. The login flow is also now state machine-based, so that's coming along!