Get rid of jitter on zoom in on graphs (#328)
This PR removes the jitter created by the Zoom Chart.js plugin. Prior to
this change, tick labels were recalculated each time the graph was
zoomed in or panned, creating a large amount of jitter. This change
calculated a set of ticks ranging from before the first datapoint to
after the most recent datapoint that is recalculated on zoom. Since this
set is static on pan, no jitter occurs.

This PR also fixes several other jitter issues, and a few miscellaneous bugs with graph rendering.

---------

Co-authored-by: nkalupahana <avishkank@gmail.com>